requirements.txt
crewai==0.86.0
crewai-tools==0.17.0
langchain==0.3.13
langchain-openai==0.2.14
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
pydantic==2.10.3
setuptools
mcp